extract base allocator and require files
