+++ lib/knapsack_pro.rb
@@ -38,7 +38,10 @@ require_relative 'knapsack_pro/test_flat_distributor'
 require_relative 'knapsack_pro/task_loader'
 require_relative 'knapsack_pro/tracker'
 require_relative 'knapsack_pro/allocator'
+require_relative 'knapsack_pro/queue_allocator'
+require_relative 'knapsack_pro/base_allocator_builder'
 require_relative 'knapsack_pro/allocator_builder'
+require_relative 'knapsack_pro/queue_allocator_builder'
 require_relative 'knapsack_pro/adapters/base_adapter'
 require_relative 'knapsack_pro/adapters/rspec_adapter'
 require_relative 'knapsack_pro/adapters/cucumber_adapter'
@@ -49,6 +52,7 @@ require_relative 'knapsack_pro/runners/rspec_runner'
 require_relative 'knapsack_pro/runners/cucumber_runner'
 require_relative 'knapsack_pro/runners/minitest_runner'
 require_relative 'knapsack_pro/runners/spinach_runner'
+require_relative 'knapsack_pro/runners/queue/base_queue_runner'
 require_relative 'knapsack_pro/runners/queue/rspec_runner'
 require_relative 'knapsack_pro/crypto/encryptor'
 require_relative 'knapsack_pro/crypto/decryptor'
