+++ lib/knapsack_pro/queue_allocator_builder.rb
@@ -1,9 +1,5 @@
 module KnapsackPro
-  class QueueAllocatorBuilder
-    def initialize(adapter_class)
-      @adapter_class = adapter_class
-    end
-
+  class QueueAllocatorBuilder < BaseAllocatorBuilder
     def allocator
       KnapsackPro::QueueAllocator.new(
         test_files: test_files,
@@ -12,30 +8,6 @@ module KnapsackPro
         ci_node_build_id: env.ci_node_build_id,
         repository_adapter: repository_adapter,
       )
-    end
-
-    def test_dir
-      test_file_pattern.split('/').first
-    end
-
-    private
-
-    attr_reader :adapter_class
-
-    def env
-      KnapsackPro::Config::Env
-    end
-
-    def repository_adapter
-      KnapsackPro::RepositoryAdapterInitiator.call
-    end
-
-    def test_file_pattern
-      TestFilePattern.call(adapter_class)
-    end
-
-    def test_files
-      KnapsackPro::TestFileFinder.call(test_file_pattern)
     end
   end
 end
